+++ pkg/domain/credentials/driven.go
@@ -4,6 +4,7 @@ package credentials
 type Storage interface {
 	SetCredentials(c Credentials) error
 	GetCredentials(url string, user string) (Credentials, error)
+	RemoveCredentials(url string, username string) error
 }
 
 // Backend ...