+++ pkg/domain/realm/driven.go
@@ -5,4 +5,5 @@ type Storage interface {
 	CreateOrUpdateRealm(r Realm) error
 	GetRealm(id string) (Realm, error)
 	ListRealms() (List, error)
+	DeleteRealm(id string) error
 }